Extract merge transform to separate file for further refactoring
